JBIDE-15317 - LiveReload script injection corrupted html encoding
Using explicitely a charset when reading local files.
Default charset used is UTF-8
Adding a combo in the server configuration editor to let the user choose her charset and override the default UTF-8 value
Fixed the "content-length" response header value
Also, the returned content is encoded with the charset provided in the request's
"accept" header. If no charset is given, the default one is used